sprint1/step2 правка метода getData, перенос обработчика нажатия Esc и тд
